Package structure reorganised
