+++ app/src/main/java/com/example/foody/data/database/RecipesDatabase.kt
@@ -1,4 +1,4 @@
-package com.example.foody
+package com.example.foody.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
